Refactor PathUtils
