+++ src/TestEngine/nunit.engine.core/Internal/TestPackageExtensions.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2019 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -23,6 +23,7 @@
 
 using System;
 using System.Collections.Generic;
+using NUnit.Engine.Utilities;
 
 namespace NUnit.Engine.Internal
 {
