+++ src/TestEngine/nunit.engine.core/Services/DomainManager.cs
@@ -34,6 +34,7 @@ using System.Security.Policy;
 using System.Security.Principal;
 using NUnit.Common;
 using NUnit.Engine.Internal;
+using NUnit.Engine.Utilities;
 
 namespace NUnit.Engine.Services
 {
