+++ src/TestEngine/nunit.engine.tests/Utilities/PathUtilTests.cs
@@ -26,7 +26,7 @@ using System.IO;
 using System.Runtime.InteropServices;
 using NUnit.Framework;
 
-namespace NUnit.Engine.Internal.Tests
+namespace NUnit.Engine.Utilities
 {
 	[TestFixture]
 	public class PathUtilTests : PathUtils
